condense icon macro
